Rename db_sema to db_semaphore
